Initial commit of extract_static_features
This function contains the 69 features derived from each frame in
isolation, i.e. the non-movement features. It currently is failing its
benchmarking against the R implementation of 50/69 values, although I
suspect 42 of these are from the Haralick features (either the Haralick
or the coccurrence matrix calculations) and 4 from the Polygon features
(arising from using the rdp library rather than the hand-rolled code as
in the original)