Mostly fixed intensity quantile disagreement
The quantiles from the real cell image are now the same as the R
version, although the test for a synthetic image is now failing (which
led to me adding in the bodge that broke it for the real image).
It looks like an issue with floating point precision, but more
investigation is needed.