Fixed intensity quantile test
The intensity quantiles are now returning the same values for both the
test image and the synthetic data.
However, this does mean the problem has been entirely resolved.
Under certain unidentified circumstances, the returned quantiles have a
slight offset (~1e-10) from the value returned by R. This is likely a
precision issue. An attempt to account for it by subtracting the
machine's floating point precision from the threshold didn't fix it.

Ultimately, the impact on the feature itself, which is the coefficient
of variation of the distance of pixels above a certain threshold, is
minimal.