Added functions to manually fill the ROI
The actual functionality of identifying the cells that lie within the
ROI hasn't been implemented yet as it relies upon having an accurate
list of the corner vertices of the polygon. This functionality has been
implemented in this commit, along with an initial attempt at removing
vertices that don't contribute to the polygon's boundary (i.e. removing
them would not affect the number of interior pixels).

This initial attempt has failed, although the corner identification
algorithm itself seems to work.