Fixed precision issue in intensity_quantiles
The np.quantile function returns floats, even for integer inputs.
Occasionally this means that the quantiles don't directly correspond to
the actual data. I.e. a quantile might be 87.000000000002, but the
actual integer pixel value corresponding to this is 87.
When comparing the pixels to this threshold (>=), this pixel won't be
included even though it is the quantile value directly.
Adding a small tolerance to the pixels when comparing against the quantiles
makes this function behave as expected.