feat: extract_features (#24)
* Added var_from_centre

* Added curvature function

* Pylint housekeeping

* Sphinx formatting

* Replaced hand-rolled euclidean distance with np implementation

* Added minimum_box function

* Added minimal polygon and interior angle calculation

* Added polygon features

* Added cooccurrence matrix calculation

* Haralick features

* Intensity quantiles

* Added wavelet transformation

* Added double_image

* Added reading ROI files (using read-roi package)

* Created test for extracting ROI mask

* Added function to create ROI mask

* Added extra test cases for mask extraction

Also corrected the fact that ROIs are given in (x,y) pairs rather than
(row, column) indices

* Added extract_subimage

* Initial commit of extract_static_features

This function contains the 69 features derived from each frame in
isolation, i.e. the non-movement features. It currently is failing its
benchmarking against the R implementation of 50/69 values, although I
suspect 42 of these are from the Haralick features (either the Haralick
or the coccurrence matrix calculations) and 4 from the Polygon features
(arising from using the rdp library rather than the hand-rolled code as
in the original)

* Corrected masks - hasn't fixed Haralick feats

Observed that the error is deriving from the cooccurrence matrix, which
is either from the cooccurrence calculation itself or the Wavelet
transformations

* Added test for Haar approx on real image

* Added more tests on real cell images

* Fixed bug in cooccurrence calculation

* Mostly fixed intensity quantile disagreement

The quantiles from the real cell image are now the same as the R
version, although the test for a synthetic image is now failing (which
led to me adding in the bodge that broke it for the real image).
It looks like an issue with floating point precision, but more
investigation is needed.

* Fixed intensity quantile test

The intensity quantiles are now returning the same values for both the
test image and the synthetic data.
However, this does mean the problem has been entirely resolved.
Under certain unidentified circumstances, the returned quantiles have a
slight offset (~1e-10) from the value returned by R. This is likely a
precision issue. An attempt to account for it by subtracting the
machine's floating point precision from the threshold didn't fix it.

Ultimately, the impact on the feature itself, which is the coefficient
of variation of the distance of pixels above a certain threshold, is
minimal.

* Fix for RDP to handle enclosed polygon

The algorithm now returns the same minimal vertex set as the original
hand-rolled code.

* Calculates static features for all cells & frames

The next step is adding the movement features

* Added movement features

* Fixed movement features and added density

* Updated benchmark after fix to minimum_box R code

* Fixes for Area and minimum box

* Added functions to manually fill the ROI

The actual functionality of identifying the cells that lie within the
ROI hasn't been implemented yet as it relies upon having an accurate
list of the corner vertices of the polygon. This functionality has been
implemented in this commit, along with an initial attempt at removing
vertices that don't contribute to the polygon's boundary (i.e. removing
them would not affect the number of interior pixels).

This initial attempt has failed, although the corner identification
algorithm itself seems to work.

* Added 3 methods for finding the interior mask

There's a custom polygon fill implementation, a point in polygon
implementation from Matplotlib, and a flood fill implementation from
scikit-image. The Matplotlib version is the quickest and has the highest
concordance with the R package.

* Fixed interior pixel discrepancy - negative flood fill

This involves filling the exterior rather than the interior, a simpler
task.
This commit also includes 2 additional interior pixel identification
algorithms: a point in polygon algorithm from skimage, and a
hand-crafted Ray Casting algorithm that casts rays from 4 directions to
identify interior pixels. Both of these two methods still have
discrepancies with the reference implementation in the R package.

* Fixed precision bug in normalise image

The transform is the same but done in a way that avoids some floating
point precision issues, i.e. whole numbers were represented as
x.99999991, resulting in downstream changes when floored in cooccurrence
matrix calculation.

* Fixed cooccurrence difference with floor tolerance

The cooccurence matrix rescales the image to integer values within [0, 10].
However due to floating point precision it can have unexpected
behaviours, for example if a number is 1.99999999996 it will be floored
to 1, when this in actuality is 2 with some imprecision. Adding a small
tolerance before the floor combats this.

* Fix corner case haralick feature bug

Occasionally, the first item in `den_1` or `den_2`, which are populated
from `pglcm`, might be NaN. When these are summed to form `den`, this
can result in a NaN value when it should be zero, thereby affecting the
Den feature, and also the DVA feature which uses Den.

* Fixed precision issue in intensity_quantiles

The np.quantile function returns floats, even for integer inputs.
Occasionally this means that the quantiles don't directly correspond to
the actual data. I.e. a quantile might be 87.000000000002, but the
actual integer pixel value corresponding to this is 87.
When comparing the pixels to this threshold (>=), this pixel won't be
included even though it is the quantile value directly.
Adding a small tolerance to the pixels when comparing against the quantiles
makes this function behave as expected.

* Removed unit tests for unsuccessful fill polygon algorithm

* Updated benchmark features following hotfix to R package

* Full integration test displays *all* discordant columns

The default `pandas.testing.assert_frame_equal` function compares 2
dataframes and raises an error if they differ. However, the error
message only names the first column in which they differ, rather than
the more useful all mismatching columns. This commit adds a custom
assert function to do this.

* Using shapely's douglas-peucker implementation

This results in an acceptable discrepancy in the reduced polygon
coordinates of 0.5% with the original R package implementation.

* Separated tests into unit & integration