feat: time series features (#47)
* summary stats, interpolation, elevation vars

* Added 1D haar approximation

* Calculate elevation vars without needing to melt

* wavelet elevation features

* Trajectory area working

* Combining time-series vars workin

The features themselves are differing from the original R implementation

* Fixed Haar and Elevation features

The only outstanding issue is the trajectoryAreas not matching up,
although I believe this is due to a bug in the original implementation.

* Fixed deprecation warning

* Updated docs

* Updating R package's features after traj area fix

During testing we identified a bug with the original R package's
implementation of trajectory area not behaving as expected when frames
were interpolated. This commit updates the target features following
this fix. The Python package is now fully concordant with the R one.

* Pylint fixes

* Error handling missing ROI files

* flake8 fixes