Save ROIs as FrameID-CellID.roi
This has been achieved by now linking the ROIs to the Frame and tracked
CellID. It produces the desired ROI files, but there are some peculiarities to sort
out. Firstly, there are 22,664 rows in the output CSV, but only 22,329 ROIs.
Secondly there are 3 cells saved in Frame 0. These 3 cells are also in
the output CSV, but it doesn't look right as the Frames generally start
at 1.