Fixed bug of too few ROIs being detected
There was a bug in the XML parsing code. Cell tracks are stored as
edges between a source cell in one frame and the linked target cell in a
subsequent frame. Previously I was saving all the target cell IDs and
the first source cell, assuming that the edges were ordered in the XML
in chronological order. However this isn't the case and so if the first
edge didn't have the first cell as its origin, then the first cell
wouldn't be included in the output.

This was resolved by a crude approach of saving all cells in all edges -
both source and targets, and then removing duplicates after. It uses
more memory but is far simpler.