Bugfix when reading in ROIs
The `roifile` library uses the `subpixel_coordinates` if a ROI file has
them, which our Trackmate ROIs do but the Phase ones don't. If they
don't then it reverts to using the `integer_coordinates`. The former are
floats (.5 of a pixel), while the latter are ints.

The package was developed around integer coordinates and so doesn't work
with the default `roifile` behaviour when reading in coordinates. This
is now done manually to ensure compatability.