Updated integration test classification
Previously the `test_classify_cells()` integration test was assessed on
the proportion of test cells that were assigned to each class,
regardless of whether these assignments were correct.
This isn't as useful a test as comparing directly on classification,
which the test now does.

The reason why the test needed updating at all is due to the move to
xgboost, the classification accuracy has increased.