fix: image processing (#138)
* Fixed non-RGB images being converted to grayscale and losing resolution

If an image was 16-bit grayscale and run through `image.convert('L')`
it would be truncated to 8-bit, i.e. any pixel value over 255 would be
clamped at 255, thereby losing a huge amount resolution.

* Bumped version number and set cellpose to v3

Cellpose v4 uses a transformer architecture which is significantly
slower on CPUs.

* Updated JPype dependency

JPype previously had a bug in 1.5.1, which has been fixed for 1.5.2

* Pinned shapely version to 2.0.7

In 2.1.0 a change was made to the RDP implementation in `simplify`,
which caused a large discrepancy (5-9%) between the polygon features
as initially calculated from the original R package. Previously this
discrepancy was within 1%.

* Updated integration test classification

Previously the `test_classify_cells()` integration test was assessed on
the proportion of test cells that were assigned to each class,
regardless of whether these assignments were correct.
This isn't as useful a test as comparing directly on classification,
which the test now does.

The reason why the test needed updating at all is due to the move to
xgboost, the classification accuracy has increased.