Manual CI replacing deprecated library
Previously [Microsoft's validation action](https://github.com/microsoft/action-python)
was used to run CI on every pull request or push to main.
In particular, it ran:

  - black
  - bandit
  - pylint
  - flake8
  - pytest

However, this action was archived in March 2025.

To replace it, I'll just run those programs manually, except for
replacing pylint and flake8 with ruff.