Added cellphe.tracking module
Previously the main `track_images` function was in `cellphe.input`,
which is not a natural home for it. By making a new module for tracking,
all the functions related to `imagej`, `trackmate` and tracking itself
can be encapsulated here, making it easier to only import these
functions when required, rather than keeping them entwined with the
cellphe core functionalty. This is a WIP and likely will necessitate a
change to `cellphe/__init__.py`