feat: cellpose and trackmate not included with default installation (#178)
* Add manual Publish action after library deprecation

Publishing to PyPI was previously achieved using Microsoft's
[Python Actions](https://github.com/microsoft/action-python), however,
this was archived in March 2025. So publishing is now handled manually
using the [PyPa's own action](https://github.com/marketplace/actions/pypi-publish).

* Black changes: indentation on module docstrings

* Manual CI replacing deprecated library

Previously [Microsoft's validation action](https://github.com/microsoft/action-python)
was used to run CI on every pull request or push to main.
In particular, it ran:

  - black
  - bandit
  - pylint
  - flake8
  - pytest

However, this action was archived in March 2025.

To replace it, I'll just run those programs manually, except for
replacing pylint and flake8 with ruff.

* ruff fixes: explicitly reexporting imports

* Create separate group for 'full' install (cellpose + trackmate)

* fix: remove XML vulnerability

* fix: add `-r` flag for ruff

* Don't install full deps for testing

* Install deps before unit tests

* Added cellphe.tracking module

Previously the main `track_images` function was in `cellphe.input`,
which is not a natural home for it. By making a new module for tracking,
all the functions related to `imagej`, `trackmate` and tracking itself
can be encapsulated here, making it easier to only import these
functions when required, rather than keeping them entwined with the
cellphe core functionalty. This is a WIP and likely will necessitate a
change to `cellphe/__init__.py`

* Moved `segment_images` to its own module

Likewise with `track_images`, this allows for cleaner separation of the
core modules from the optional ones.

* Skip tests that require optional deps

Also mark them as such so can explicitly target them

* Hacky way to import optional modules at import time

Create a dummy function that wraps the importing logic, so calling code
can still access the functions with the top-level import, they just
aren't imported at module first import.

* Core unit & integration tests run successfully

* Tests (both unit & integration) of full deps working

* Run tests on full dependencies

* Warn when trying to segment or track without full deps

If the user tries to run `segment_images` or `track_images` without the
[full] dependencies, a warning message will be displayed.

* Updated READMEs with instructions to install full deps

I.e. `pip install cellphe[full]` to use segmentation and tracking.

* Fix import bugs with tracking & segmenting

Also confirm README notebook still runs

* Bump version number

* Removed unused import (ruff)

* Removed Spark from pyproject.toml

This was leftover from the template

* Fixed pytest in CI

`uvx pytest` uses a general `pytest` install, not the one specific to
this project which has the `pytest-cov` package installed. This means
some of the flags fail. The solution is to force the local `pytest` is
run instead with `uv run pytest ...`

* Fixed typo in CI